proyecto-03/requirements.txt
RPi.GPIO>=0.7.1
gpiozero>=2.0
opencv-python>=4.5.0
Pillow>=9.0.0
numpy>=1.21.0
edge-impulse-linux>=1.0.0
